requirements.txt
aniso8601==3.0.2
astroid==2.0.4
certifi==2018.10.15
chardet==3.0.4
Click==7.0
colorama==0.3.9
coverage==4.5.1
coveralls==1.5.1
docopt==0.6.2
Flask==1.0.2
Flask-RESTful==0.3.6
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
flask-expects-json==1.3.1
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
nose==1.3.7
postgres==2.2.2
psycopg2-binary==2.7.5
PyJWT==1.6.4
pylint==2.1.1
pytz==2018.5
requests==2.20.0
six==1.11.0
urllib3==1.24
virtualenv==16.0.0
Werkzeug==0.14.1
wrapt==1.10.11
